+++ lib/mustache.rb
@@ -1,5 +1,6 @@
 require 'mustache/template'
 require 'mustache/context'
+require 'mustache/settings'
 
 # Mustache is the base class from which your Mustache subclasses
 # should inherit (though it can be used on its own).
@@ -39,7 +40,7 @@ require 'mustache/context'
 #
 # * template_file
 #
-# You can tell Mustache exactly which template to us with this
+# You can tell Mustache exactly which template to use with this
 # setting. It can be a relative or absolute path.
 #
 # * template
@@ -70,19 +71,83 @@ require 'mustache/context'
 # for files containing view classes when using the `view_class` method.
 #
 class Mustache
-  # Helper method for quickly instantiating and rendering a view.
+
+  #
+  # Public API
+  #
+
+  # Instantiates an instance of this class and calls `render` with
+  # the passed args.
+  #
+  # Returns a rendered String version of a template
   def self.render(*args)
     new.render(*args)
   end
 
-  # Alias for `render`
-  def self.to_html(*args)
-    render(*args)
+  class << self
+    alias_method :to_html, :render
+    alias_method :to_text, :render
   end
 
-  # Alias for `render`
-  def self.to_text(*args)
-    render(*args)
+  # Parses our fancy pants template file and returns normal file with
+  # all special {{tags}} and {{#sections}}replaced{{/sections}}.
+  #
+  # data - A String template or a Hash context. If a Hash is given,
+  #        we'll try to figure out the template from the class.
+  #  ctx - A Hash context if `data` is a String template.
+  #
+  # Examples
+  #
+  #   @view.render("Hi {{thing}}!", :thing => :world)
+  #
+  #   View.template = "Hi {{thing}}!"
+  #   @view = View.new
+  #   @view.render(:thing => :world)
+  #
+  # Returns a rendered String version of a template
+  def render(data = template, ctx = {})
+    if data.is_a? Hash
+      ctx = data
+      tpl = templateify(template)
+    elsif data.is_a? Symbol
+      self.template_name = data
+      tpl = templateify(template)
+    else
+      tpl = templateify(data)
+    end
+
+    return tpl.render(context) if ctx == {}
+
+    begin
+      context.push(ctx)
+      tpl.render(context)
+    ensure
+      context.pop
+    end
+  end
+
+  alias_method :to_html, :render
+  alias_method :to_text, :render
+
+  # Context accessors.
+  #
+  # view = Mustache.new
+  # view[:name] = "Jon"
+  # view.template = "Hi, {{name}}!"
+  # view.render # => "Hi, Jon!"
+  def [](key)
+    context[key.to_sym]
+  end
+
+  def []=(key, value)
+    context[key.to_sym] = value
+  end
+
+  # A helper method which gives access to the context at a given time.
+  # Kind of a hack for now, but useful when you're in an iterating section
+  # and want access to the hash currently being iterated over.
+  def context
+    @context ||= Context.new(self)
   end
 
   # Given a file name and an optional context, attempts to load and
@@ -113,91 +178,23 @@ class Mustache
     self.class.partial(name)
   end
 
-  # The template path informs your Mustache subclass where to look for its
-  # corresponding template. By default it's the current directory (".")
-  def self.template_path
-    @template_path ||= inheritable_config_for :template_path, '.'
+  # Override this to provide custom escaping.
+  #
+  # class PersonView < Mustache
+  #   def escapeHTML(str)
+  #     my_html_escape_method(str)
+  #   end
+  # end
+  #
+  # Returns a String
+  def escapeHTML(str)
+    CGI.escapeHTML(str)
   end
 
-  def self.template_path=(path)
-    @template_path = File.expand_path(path)
-    @template = nil
-  end
 
-  # Alias for `template_path`
-  def self.path
-    template_path
-  end
-
-  # Alias for `template_path`
-  def self.path=(path)
-    self.template_path = path
-  end
-
-  # A Mustache template's default extension is 'mustache'
-  def self.template_extension
-    @template_extension ||= inheritable_config_for :template_extension, 'mustache'
-  end
-
-  def self.template_extension=(template_extension)
-    @template_extension = template_extension
-    @template = nil
-  end
-
-  # The template name is the Mustache template file without any
-  # extension or other information. Defaults to `class_name`.
-  def self.template_name
-    @template_name || underscore
-  end
-
-  def self.template_name=(template_name)
-    @template_name = template_name
-    @template = nil
-  end
-
-  # The template file is the absolute path of the file Mustache will
-  # use as its template. By default it's ./class_name.mustache
-  def self.template_file
-    @template_file || "#{path}/#{template_name}.#{template_extension}"
-  end
-
-  def self.template_file=(template_file)
-    @template_file = template_file
-    @template = nil
-  end
-
-  # The template is the actual string Mustache uses as its template.
-  # There is a bit of magic here: what we get back is actually a
-  # Mustache::Template object here, but you can still safely use
-  # `template=` with a string.
-  def self.template
-    @template ||= templateify(File.read(template_file))
-  end
-
-  def self.template=(template)
-    @template = templateify(template)
-  end
-
-  # The constant under which Mustache will look for views. By default it's
-  # `Object`, but it might be nice to set it to something like `Hurl::Views` if
-  # your app's main namespace is `Hurl`.
-  def self.view_namespace
-    @view_namespace ||= inheritable_config_for(:view_namespace, Object)
-  end
-
-  def self.view_namespace=(namespace)
-    @view_namespace = namespace
-  end
-
-  # Mustache searches the view path for .rb files to require when asked to find a
-  # view class. Defaults to "."
-  def self.view_path
-    @view_path ||= inheritable_config_for(:view_path, '.')
-  end
-
-  def self.view_path=(path)
-    @view_path = path
-  end
+  #
+  # Private API
+  #
 
   # When given a symbol or string representing a class, will try to produce an
   # appropriate view class.
@@ -215,38 +212,33 @@ class Mustache
     end
 
     file_name = underscore(name)
-    namespace = view_namespace
+    name = "#{view_namespace}::#{name}"
 
-    if namespace.const_defined?(:Views) && namespace::Views.const_defined?(name)
-      namespace::Views.const_get(name)
-    elsif namespace.const_defined?(name)
-      namespace.const_get(name)
+    if const = const_get!(name)
+      const
     elsif File.exists?(file = "#{view_path}/#{file_name}.rb")
       require "#{file}".chomp('.rb')
-      if namespace.const_defined?(:Views)
-        namespace::Views.const_get(name)
-      else
-        namespace.const_get(name)
-      end
+      const_get!(name) || Mustache
     else
       Mustache
     end
-  rescue NameError
-    Mustache
   end
 
-  # Should an exception be raised when we cannot find a corresponding method
-  # or key in the current context? By default this is false to emulate ctemplate's
-  # behavior, but it may be useful to enable when debugging or developing.
+  # Supercharged version of Module#const_get.
   #
-  # If set to true and there is a context miss, `Mustache::ContextMiss` will
-  # be raised.
-  def self.raise_on_context_miss?
-    @raise_on_context_miss
-  end
-
-  def self.raise_on_context_miss=(boolean)
-    @raise_on_context_miss = boolean
+  # Always searches under Object and can find constants by their full name,
+  #   e.g. Mustache::Views::Index
+  #
+  # name - The full constant name to find.
+  #
+  # Returns the constant if found
+  # Returns nil if nothing is found
+  def self.const_get!(name)
+    name.split('::').inject(Object) do |klass, cname|
+      klass.const_get(cname)
+    end
+  rescue NameError
+    nil
   end
 
   # Has this template already been compiled? Compilation is somewhat
@@ -261,21 +253,28 @@ class Mustache
   end
 
   # template_partial => TemplatePartial
+  # template/partial => Template::Partial
   def self.classify(underscored)
-    underscored.split(/[-_]/).map do |part|
-      part[0] = part[0].chr.upcase; part
-    end.join
+    underscored.split('/').map do |namespace|
+      namespace.split(/[-_]/).map do |part|
+        part[0] = part[0].chr.upcase; part
+      end.join
+    end.join('::')
   end
 
-  # TemplatePartial => template_partial
+  #   TemplatePartial => template_partial
+  # Template::Partial => template/partial
   # Takes a string but defaults to using the current class' name.
   def self.underscore(classified = name)
     classified = name if classified.to_s.empty?
     classified = superclass.name if classified.to_s.empty?
 
-    string = classified.dup.split('::').last
-    string[0] = string[0].chr.downcase
-    string.gsub(/[A-Z]/) { |s| "_#{s.downcase}"}
+    string = classified.dup.split("#{view_namespace}::").last
+
+    string.split('::').map do |part|
+      part[0] = part[0].chr.downcase
+      part.gsub(/[A-Z]/) { |s| "_#{s.downcase}"}
+    end.join('/')
   end
 
   # Turns a string into a Mustache::Template. If passed a Template,
@@ -288,7 +287,11 @@ class Mustache
     end
   end
 
-  # Return the value of the configuration setting on the superclass, or return 
+  def templateify(obj)
+    self.class.templateify(obj)
+  end
+
+  # Return the value of the configuration setting on the superclass, or return
   # the default.
   #
   # attr_name - Symbol name of the attribute.  It should match the instance variable.
@@ -298,61 +301,4 @@ class Mustache
   def self.inheritable_config_for(attr_name, default)
     superclass.respond_to?(attr_name) ? superclass.send(attr_name) : default
   end
-
-  def templateify(obj)
-    self.class.templateify(obj)
-  end
-
-  # The template can be set at the instance level.
-  def template
-    @template ||= self.class.template
-  end
-
-  def template=(template)
-    @template = templateify(template)
-  end
-
-  # Instance level version of `Mustache.raise_on_context_miss?`
-  def raise_on_context_miss?
-    self.class.raise_on_context_miss? || @raise_on_context_miss
-  end
-  attr_writer :raise_on_context_miss
-
-  # A helper method which gives access to the context at a given time.
-  # Kind of a hack for now, but useful when you're in an iterating section
-  # and want access to the hash currently being iterated over.
-  def context
-    @context ||= Context.new(self)
-  end
-
-  # Context accessors.
-  #
-  # view = Mustache.new
-  # view[:name] = "Jon"
-  # view.template = "Hi, {{name}}!"
-  # view.render # => "Hi, Jon!"
-  def [](key)
-    context[key.to_sym]
-  end
-
-  def []=(key, value)
-    context[key.to_sym] = value
-  end
-
-  # Parses our fancy pants template file and returns normal file with
-  # all special {{tags}} and {{#sections}}replaced{{/sections}}.
-  def render(data = template, ctx = {})
-    tpl = templateify(data)
-
-    return tpl.render(context) if ctx == {}
-
-    begin
-      context.push(ctx)
-      tpl.render(context)
-    ensure
-      context.pop
-    end
-  end
-  alias_method :to_html, :render
-  alias_method :to_text, :render
 end