+++ lib/mustache/generator.rb
@@ -90,7 +90,7 @@ class Mustache
 
     # Callback fired when the compiler finds a section token. We're
     # passed the section name and the array of tokens.
-    def on_section(name, content)
+    def on_section(name, content, raw, delims)
       # Convert the tokenized content of this section into a Ruby
       # string we can use.
       code = compile(content)
@@ -98,14 +98,20 @@ class Mustache
       # Compile the Ruby for this section now that we know what's
       # inside the section.
       ev(<<-compiled)
-      if v = ctx[#{name.to_sym.inspect}]
+      if v = #{compile!(name)}
         if v == true
           #{code}
         elsif v.is_a?(Proc)
-          v.call(#{code})
+          t = Mustache::Template.new(v.call(#{raw.inspect}).to_s)
+          def t.tokens(src=@source)
+            p = Parser.new
+            p.otag, p.ctag = #{delims.inspect}
+            p.compile(src)
+          end
+          t.render(ctx.dup)
         else
           # Shortcut when passed non-array
-          v = [v] if v.respond_to?(:has_key?) || !v.respond_to?(:map)
+          v = [v] unless v.is_a?(Array) || defined?(Enumerator) && v.is_a?(Enumerator)
 
           v.map { |h| ctx.push(h); r = #{code}; ctx.pop; r }.join
         end
@@ -115,7 +121,7 @@ class Mustache
 
     # Fired when we find an inverted section. Just like `on_section`,
     # we're passed the inverted section name and the array of tokens.
-    def on_inverted_section(name, content)
+    def on_inverted_section(name, content, raw, _)
       # Convert the tokenized content of this section into a Ruby
       # string we can use.
       code = compile(content)
@@ -123,7 +129,7 @@ class Mustache
       # Compile the Ruby for this inverted section now that we know
       # what's inside.
       ev(<<-compiled)
-      v = ctx[#{name.to_sym.inspect}]
+      v = #{compile!(name)}
       if v.nil? || v == false || v.respond_to?(:empty?) && v.empty?
         #{code}
       end
@@ -133,18 +139,47 @@ class Mustache
     # Fired when the compiler finds a partial. We want to return code
     # which calls a partial at runtime instead of expanding and
     # including the partial's body to allow for recursive partials.
-    def on_partial(name)
-      ev("ctx.partial(#{name.to_sym.inspect})")
+    def on_partial(name, indentation)
+      ev("ctx.partial(#{name.to_sym.inspect}, #{indentation.inspect})")
     end
 
     # An unescaped tag.
     def on_utag(name)
-      ev("ctx[#{name.to_sym.inspect}]")
+      ev(<<-compiled)
+        v = #{compile!(name)}
+        if v.is_a?(Proc)
+          v = Mustache::Template.new(v.call.to_s).render(ctx.dup)
+        end
+        v.to_s
+      compiled
     end
 
     # An escaped tag.
     def on_etag(name)
-      ev("CGI.escapeHTML(ctx[#{name.to_sym.inspect}].to_s)")
+      ev(<<-compiled)
+        v = #{compile!(name)}
+        if v.is_a?(Proc)
+          v = Mustache::Template.new(v.call.to_s).render(ctx.dup)
+        end
+        ctx.escapeHTML(v.to_s)
+      compiled
+    end
+
+    def on_fetch(names)
+      names = names.map { |n| n.to_sym }
+
+      if names.length == 0
+        "ctx[:to_s]"
+      elsif names.length == 1
+        "ctx[#{names.first.to_sym.inspect}]"
+      else
+        initial, *rest = names
+        <<-compiled
+          #{rest.inspect}.inject(ctx[#{initial.inspect}]) { |value, key|
+            value && ctx.find(value, key)
+          }
+        compiled
+      end
     end
 
     # An interpolation-friendly version of a string, for use within a