+++ test/parser_test.rb
@@ -22,32 +22,56 @@ EOF
 
     expected = [:multi,
       [:static, "<h1>"],
-      [:mustache, :etag, "header"],
+      [:mustache, :etag, [:mustache, :fetch, ["header"]]],
       [:static, "</h1>\n"],
       [:mustache,
         :section,
-        "items",
+        [:mustache, :fetch, ["items"]],
         [:multi,
           [:mustache,
             :section,
-            "first",
+            [:mustache, :fetch, ["first"]],
             [:multi,
-              [:static, "<li><strong>"],
-              [:mustache, :etag, "name"],
-              [:static, "</strong></li>\n"]]],
+              [:static, "  <li><strong>"],
+              [:mustache, :etag, [:mustache, :fetch, ["name"]]],
+              [:static, "</strong></li>\n"]],
+            %Q'  <li><strong>{{name}}</strong></li>\n',
+            %w[{{ }}]],
           [:mustache,
             :section,
-            "link",
+            [:mustache, :fetch, ["link"]],
             [:multi,
-              [:static, "<li><a href=\""],
-              [:mustache, :etag, "url"],
+              [:static, "  <li><a href=\""],
+              [:mustache, :etag, [:mustache, :fetch, ["url"]]],
               [:static, "\">"],
-              [:mustache, :etag, "name"],
-              [:static, "</a></li>\n"]]]]],
+              [:mustache, :etag, [:mustache, :fetch, ["name"]]],
+              [:static, "</a></li>\n"]],
+            %Q'  <li><a href="{{url}}">{{name}}</a></li>\n',
+            %w[{{ }}]]],
+        %Q'{{#first}}\n  <li><strong>{{name}}</strong></li>\n{{/first}}\n{{#link}}\n  <li><a href="{{url}}">{{name}}</a></li>\n{{/link}}\n',
+        %w[{{ }}]],
+      [:static, "\n"],
       [:mustache,
         :section,
-        "empty",
-        [:multi, [:static, "<p>The list is empty.</p>\n"]]]]
+        [:mustache, :fetch, ["empty"]],
+        [:multi, [:static, "<p>The list is empty.</p>\n"]],
+        %Q'<p>The list is empty.</p>\n',
+        %w[{{ }}]]]
+
+    assert_equal expected, tokens
+  end
+
+  def test_raw_content_and_whitespace
+    lexer = Mustache::Parser.new
+    tokens = lexer.compile("{{#list}}\t{{/list}}")
+
+    expected = [:multi,
+      [:mustache,
+        :section,
+        [:mustache, :fetch, ["list"]],
+        [:multi, [:static, "\t"]],
+        "\t",
+        %w[{{ }}]]]
 
     assert_equal expected, tokens
   end