+++ HISTORY.md
@@ -1,3 +1,50 @@
+## 0.99.4 (2011-05-26)
+
+* Reverted automatically attr_accessors in Sinatra. Too much magic.
+
+## 0.99.3 (2011-03-06)
+
+* Fixed method_missing support
+* Fixed whitespace errors
+
+## 0.99.2 (2011-03-06)
+
+* Added some HTML5 tags to tpl-mode.el
+* Bugfix: {{#sections}} and whitespace
+
+## 0.99.1 (2011-03-01)
+
+* Bugfix: Strings used in sections no longer enumerate.
+
+## 0.99.0 (2011-02-28)
+
+* Unary methods of Mustache view can now be used in sections as procs.
+* Added instance-level versions of all view settings.
+* Emacs mode: Support for dot notation
+* Sinatra: Support for `mustache "views/in/sub/directories"`
+
+## 0.98.0 (2011-02-24)
+
+* Dot Notation: {{person.name}}
+* Implicit Iterators: {{# people}} {{.}} {{/ people}}
+* Automagic attr_readers for ivars you set in Sinatra routes.
+* Ability to customize escaping by overriding Mustache#escapeHTML
+
+## 0.13.0 (2011-02-23)
+
+* Mustache.render accepts a symbol to specify a template name.
+* Bugfix: Struct can be used as sections again
+
+## 0.12.1 (2011-02-22)
+
+* Bugfixes!
+
+## 0.12.0 (2010-12-10)
+
+* Passing the full draft Mustache spec.
+* Instance render: `@view.render(:thing => :world)`
+* Sinatra bugfix: 'layout not found' fix
+
 ## 0.11.2 (2010-05-13)
 
 * Sinatra bugfix: `mustache :"TEMPLATE.atom"`